refactor: centralize mock classes, timeout helpers
