+++ tests/test_app.js
@@ -1,164 +1,7 @@
 import { assert, spyOn } from "./utils.js";
 import { Card } from "../src/deck.js";
 import { App } from "../src/app.js";
-import { Game } from "../src/game.js";
-import { TestAI as _TestAI } from "./test_ai.js";
-
-// Mock objects
-const mockDeck = {
-  cards: [
-    new Card("3", "♠"),
-    new Card("4", "♦"),
-    new Card("5", "♣"),
-    new Card("6", "♥"),
-    new Card("7", "♠"),
-    new Card("8", "♦"),
-    new Card("9", "♣"),
-    new Card("10", "♥"),
-    new Card("J", "♠"),
-    new Card("Q", "♦"),
-    new Card("K", "♣"),
-    new Card("A", "♥"),
-    new Card("2", "♠"),
-    new Card("3", "♦"),
-    new Card("4", "♣"),
-    new Card("5", "♥"),
-    new Card("6", "♠"),
-    new Card("7", "♦"),
-    new Card("8", "♣"),
-    new Card("9", "♥"),
-    new Card("10", "♠"),
-    new Card("J", "♦"),
-    new Card("Q", "♣"),
-    new Card("K", "♥"),
-    new Card("A", "♠"),
-    new Card("2", "♦"),
-  ],
-  deal: () => [
-    [
-      new Card("3", "♠"),
-      new Card("4", "♦"),
-      new Card("5", "♣"),
-      new Card("6", "♥"),
-      new Card("7", "♠"),
-      new Card("8", "♦"),
-      new Card("9", "♣"),
-      new Card("10", "♥"),
-      new Card("J", "♠"),
-      new Card("Q", "♦"),
-      new Card("K", "♣"),
-      new Card("A", "♥"),
-      new Card("2", "♠"),
-    ],
-    [
-      new Card("3", "♦"),
-      new Card("4", "♣"),
-      new Card("5", "♥"),
-      new Card("6", "♠"),
-      new Card("7", "♦"),
-      new Card("8", "♣"),
-      new Card("9", "♥"),
-      new Card("10", "♠"),
-      new Card("J", "♦"),
-      new Card("Q", "♣"),
-      new Card("K", "♥"),
-      new Card("A", "♠"),
-      new Card("2", "♦"),
-    ],
-  ],
-  shuffle: () => {},
-};
-
-// Mock Game class by extending the actual Game class
-class MockGame extends Game {
-  constructor() {
-    super();
-    this.resetCalled = false;
-    this.findStartingPlayerCalled = false;
-    this.playCardsCalled = false;
-    this.passTurnCalled = false;
-    this.startCalled = false;
-  }
-
-  reset() {
-    this.resetCalled = true;
-    super.reset();
-  }
-  findStartingPlayer(hands) {
-    this.findStartingPlayerCalled = true;
-    return super.findStartingPlayer(hands);
-  }
-  playCards() {
-    this.playCardsCalled = true;
-    super.playCards();
-  }
-  passTurn() {
-    this.passTurnCalled = true;
-    return super.passTurn();
-  }
-  start(deck) {
-    this.startCalled = true;
-    super.start(deck);
-  }
-}
-
-export class MockUI {
-  constructor(game) {
-    this.game = game;
-    this.initCalled = false;
-    this.renderCalled = false;
-    this.playButton = {
-      addEventListener: (event, handler) => {
-        this.playButton.handler = handler;
-      },
-    };
-    this.passButton = {
-      addEventListener: (event, handler) => {
-        this.passButton.handler = handler;
-      },
-    };
-    this.newGameButton = {
-      addEventListener: (event, handler) => {
-        this.newGameButton.handler = handler;
-      },
-    };
-    this.gameMessages = { textContent: "", classList: { add: () => {}, remove: () => {} } };
-    this.playersHands = { innerHTML: "", appendChild: () => {} };
-    this.playArea = { innerHTML: "", appendChild: () => {} };
-    this.gameContent = { innerHTML: "", appendChild: () => {} };
-  }
-  init(game) {
-    this.game = game;
-    this.initCalled = true;
-  }
-  render() {
-    this.renderCalled = true;
-  }
-  displayMessage(message) {
-    this.gameMessages.textContent = message;
-  }
-  clearMessage() {}
-  createCardElement() {
-    return { classList: { add: () => {}, remove: () => {} }, dataset: {}, addEventListener: () => {} };
-  }
-  renderCardsContainer() {}
-  renderPlayArea() {}
-  renderPlayerHand() {}
-  renderPlayerHands() {}
-  updateButtonStates() {}
-  renderSelectedCards() {}
-}
-
-class MockAI extends _TestAI {
-  constructor(game, move = []) {
-    super(game, move);
-    this.takeTurnCalled = false;
-  }
-  takeTurn(hand, playPile, currentTurn, allPlayerHands) {
-    this.takeTurnCalled = true;
-    return super.takeTurn(hand, playPile, currentTurn, allPlayerHands);
-  }
-}
+import { MockDeck, MockGame, MockUI, MockAI } from "./mocks.js";
 
 // Test cases
 
@@ -166,9 +9,9 @@ function test_handleAITurn_playsCardsWhenMoveIsAvailable() {
   const game = new MockGame();
   // Pass the AI's move directly to the MockAI constructor
   const app = new App(
-    mockDeck,
+    MockDeck,
     game,
-    new MockAI(game, [new Card("3", "♠")]), // AI will play 3♠,
+    new MockAI(game, [new Card("3", "♠")]), // AI will play 3♠
     new MockUI(game)
   );
 
@@ -206,7 +49,7 @@ function test_handleAITurn_passesTurnWhenNoMoveIsAvailable() {
   const game = new MockGame();
   // Pass the AI's move directly to the MockAI constructor
   const app = new App(
-    mockDeck,
+    MockDeck,
     game,
     new MockAI(game, []), // AI will pass
     new MockUI(game)
@@ -222,7 +65,7 @@ function test_handleAITurn_passesTurnWhenNoMoveIsAvailable() {
   };
 
   // Mock setTimeout to prevent immediate execution of nextTurn during init
-  app.init((handler, delay) => { /* do nothing */ });
+  app.init((handler, delay) => {});
 
   app.game.gameState.currentPlayer = 1; // Set current player to AI
   app.nextTurn = originalNextTurn; // Restore nextTurn
@@ -244,7 +87,7 @@ function test_handleAITurn_passesTurnWhenNoMoveIsAvailable() {
 
 function test_handleHumanPlay_callsPlayButtonClickAndNextTurn() {
   const game = new MockGame();
-  const app = new App(mockDeck, game, new MockAI(game), new MockUI(game));
+  const app = new App(MockDeck, game, new MockAI(game), new MockUI(game));
   app.init();
   app.game.gameState.currentPlayer = 0; // the human player
 
@@ -262,6 +105,7 @@ function test_handleHumanPass_callsPassButtonClickAndNextTurn() {
   const game = new MockGame();
   const app = new App(mockDeck, game, new MockAI(game), new MockUI(game));
 
+  const app = new App(MockDeck, game, new MockAI(game), new MockUI(game));
   app.init();
   app.game.gameState.currentPlayer = 0; // the human player
 
@@ -278,10 +122,28 @@ function test_handleHumanPass_callsPassButtonClickAndNextTurn() {
   nextTurnSpy.restore();
 }
 
+function test_init_initializesGameStateCorrectly() {
+  const game = new MockGame();
+
+  const app = new App(MockDeck, game, new MockAI(game), new MockUI(game));
+
+  app.init();
+
+  assert(app.game.resetCalled, "game.reset should be called");
+  assert(app.game.startCalled, "game.start should be called");
+  assert(app.game.gameState.players.length === 2, "init should initialize 2 players");
+  assert(app.game.gameState.playerTypes[0] === "human", "First player should be human");
+  assert(app.game.gameState.playerTypes[1] === "ai", "Second player should be AI");
+  assert(app.game.findStartingPlayerCalled, "findStartingPlayer should be called");
+  assert(app.ui.initCalled, "ui.init should be called");
+  assert(app.ui.renderCalled, "ui.render should be called");
+}
+
 function test_nextTurn_callsHandleAITurnForAIPlayer() {
   const game = new MockGame();
   const app = new App(mockDeck, game, new MockAI(game, [new Card("3", "♦")]), new MockUI(game));
 
+  const app = new App(MockDeck, game, new MockAI(game, [new Card("3", "♦")]), new MockUI(game));
   const handleAITurnSpy = spyOn(app, "handleAITurn");
 
   // Set up playPile and currentPlay for the AI to respond to
@@ -301,7 +163,7 @@ function test_nextTurn_callsHandleAITurnForAIPlayer() {
 
 function test_nextTurn_doesNotCallHandleAITurnForHumanPlayer() {
   const game = new MockGame();
-  const app = new App(mockDeck, game, new MockAI(game), new MockUI(game));
+  const app = new App(MockDeck, game, new MockAI(game), new MockUI(game));
   const handleAITurnSpy = spyOn(app, "handleAITurn");
 
   app.init();
@@ -312,22 +174,6 @@ function test_nextTurn_doesNotCallHandleAITurnForHumanPlayer() {
   assert(!handleAITurnSpy.called, "handleAITurn should not be called");
 
   handleAITurnSpy.restore();
-}
-
-function test_startGame_initializesGameStateCorrectly() {
-  const game = new MockGame();
-  const app = new App(mockDeck, game, new MockAI(game), new MockUI(game));
-
-  app.init();
-
-  assert(app.game.resetCalled, "game.reset should be called");
-  assert(app.game.startCalled, "game.start should be called");
-  assert(app.game.gameState.players.length === 2, "init should initialize 2 players");
-  assert(app.game.gameState.playerTypes[0] === "human", "First player should be human");
-  assert(app.game.gameState.playerTypes[1] === "ai", "Second player should be AI");
-  assert(app.game.findStartingPlayerCalled, "findStartingPlayer should be called");
-  assert(app.ui.initCalled, "ui.init should be called");
-  assert(app.ui.renderCalled, "ui.render should be called");
 }
 
 function test_spyOn_MockAI_takeTurn() {
@@ -346,8 +192,8 @@ export const appTests = [
   test_handleAITurn_passesTurnWhenNoMoveIsAvailable,
   test_handleHumanPlay_callsPlayButtonClickAndNextTurn,
   test_handleHumanPass_callsPassButtonClickAndNextTurn,
+  test_init_initializesGameStateCorrectly,
   test_nextTurn_callsHandleAITurnForAIPlayer,
   test_nextTurn_doesNotCallHandleAITurnForHumanPlayer,
-  test_startGame_initializesGameStateCorrectly,
   test_spyOn_MockAI_takeTurn,
 ];
